+++ test/barcode_scanner_plugin_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:barcode_scanner_plugin/barcode_scanner_plugin.dart';
 import 'package:barcode_scanner_plugin/barcode_scanner_plugin_platform_interface.dart';
 import 'package:barcode_scanner_plugin/barcode_scanner_plugin_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -8,12 +7,10 @@ class MockBarcodeScannerPluginPlatform
     with MockPlatformInterfaceMixin
     implements BarcodeScannerPluginPlatform {
   @override
-  // TODO: implement barcodeStream
   Stream<String> get barcodeStream => throw UnimplementedError();
 
   @override
   Future<String?> startCameraScan() {
-    // TODO: implement startCameraScan
     throw UnimplementedError();
   }
 }